requirements.txt
pytz
requests
beautifulsoup4
feedparser
python-dateutil